backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
pydantic==2.9.0
pydantic-settings==2.5.0
python-dotenv==1.0.1
httpx==0.27.0
anthropic==0.34.0
supabase==2.7.0
python-jose[cryptography]==3.3.0
python-multipart==0.0.9
passlib[bcrypt]==1.7.4
alembic==1.13.0
sqlalchemy==2.0.32
structlog==24.4.0
